catalogue: bug: Fix product ID type comparison
The endpoint to get a product by ID was not working correctly as the
`product_id` was stored as a string in the products.json file and the
method was searching for it as a string. This was causing an incorrect
product not found error. Casting the int as a string so the product can
be found correctly in the products dictionary.